+++ src/test/java/uk/q3c/krail/core/i18n/DefaultI18NProcessorTest.java
@@ -33,7 +33,7 @@ import uk.q3c.krail.core.guice.vsscope.VaadinSessionScopeModule;
 import uk.q3c.krail.core.vaadin.MockVaadinSession;
 import uk.q3c.krail.eventbus.mbassador.EventBusModule;
 import uk.q3c.krail.i18n.CurrentLocale;
-import uk.q3c.krail.option.mock.MockOption;
+import uk.q3c.krail.option.Option;
 import uk.q3c.krail.option.mock.TestOptionModule;
 import uk.q3c.krail.persist.inmemory.InMemoryModule;
 import uk.q3c.krail.persist.inmemory.InMemoryOptionStore;
@@ -58,7 +58,7 @@ public class DefaultI18NProcessorTest {
     CurrentLocale currentLocale;
 
     @Inject
-    MockOption option;
+    Option option;
 
     @Inject
     DefaultI18NProcessor processor;