+++ src/test/java/uk/q3c/krail/core/navigate/DefaultNavigatorTest.java
@@ -36,6 +36,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import uk.q3c.krail.config.ConfigurationFileModule;
 import uk.q3c.krail.core.config.KrailApplicationConfigurationModule;
 import uk.q3c.krail.core.eventbus.UIBus;
 import uk.q3c.krail.core.eventbus.UIBusProvider;
@@ -73,7 +74,6 @@ import uk.q3c.krail.eventbus.BusMessage;
 import uk.q3c.krail.eventbus.SubscribeTo;
 import uk.q3c.krail.eventbus.mbassador.EventBusModule;
 import uk.q3c.krail.i18n.test.TestI18NModule;
-import uk.q3c.krail.option.mock.MockOption;
 import uk.q3c.krail.option.mock.TestOptionModule;
 import uk.q3c.krail.persist.inmemory.InMemoryModule;
 import uk.q3c.krail.testutil.guice.uiscope.TestUIScopeModule;
@@ -95,7 +95,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MycilaJunitRunner.class)
-@GuiceContext({VaadinSessionScopeModule.class, TestI18NModule.class, InMemoryModule.class, TestOptionModule.class, VaadinEventBusModule.class,
+@GuiceContext({VaadinSessionScopeModule.class, TestI18NModule.class, InMemoryModule.class, TestOptionModule.class, VaadinEventBusModule.class, ConfigurationFileModule.class,
         TestUIScopeModule.class, ServletEnvironmentModule.class, SerializationSupportModule.class, EventBusModule.class, SitemapModule.class, UtilModule.class, UtilsModule.class, KrailApplicationConfigurationModule.class})
 public class DefaultNavigatorTest {
 
@@ -107,8 +107,7 @@ public class DefaultNavigatorTest {
     BeforeViewChangeBusMessage event;
     @Mock
     UserStatusChangeSource source;
-    @Inject
-    MockOption option;
+
     @Inject
     DefaultViewChangeRule defaultViewChangeRule;
     @Mock