+++ src/test/java/uk/q3c/krail/core/ui/ScopedUITest.java
@@ -31,7 +31,7 @@ import org.mockito.Matchers;
 import org.mockito.Mock;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
-import uk.q3c.krail.config.ConfigurationException;
+import uk.q3c.krail.core.ConfigurationException;
 import uk.q3c.krail.core.eventbus.UIBusProvider;
 import uk.q3c.krail.core.guice.uiscope.UIKey;
 import uk.q3c.krail.core.guice.uiscope.UIScope;